Update patch set 1
Patch Set 1:

Just wondering, what's the status of this patch? Should it be merged sometime soon? Was playing with bringing up another qcom device (d2vzw) and I keep having to cherry pick this, would be nice if it was merged (and it's always nice to fix typos anyway).

Patch-set: 1