Update patch set 1
Patch Set 1:

Main thing is that ALL devices need to make the switchover at once.  That's pretty difficult to coordinate right now, pulser is working on setting up a maintainer mailing list so we can better handle announcements and coordination.  (Since a lot of people, including admittedly myself, don't necessarily notice announcements in the topic of #omni)

First order of business at the moment is to switch media/display-caf to the new qcom-4.4 branches.  This seems ready (I think all devices have tested it now except for possibly jflte) and will probably happen this weekend.

Next is to switch audio-caf to the new branch.  This needs a few fixups on some devices first.

Next I think is switching over this little patch.  :)

Patch-set: 1