Update patch set 2
Patch Set 2:

As far as a maintainers hangout - we already have IRC for "chat-style" coordination.  Hangouts just fragments the communications medium since it serves the exact same use case as IRC.

A mailing list is for items that aren't suitable for Hangouts - e.g. where the discussion requires people to be more verbose than an IRC chat.

Private G+ is a non-starter for various reasons:
1)  There's no way I know of to have a group that is "read-only" for the general public but postable by approved members.  A fully private group is a non-starter, most of the original Omni team founded the project partly in reaction to CM's lack of transparency.
2)  Certain key people refuse to use G+ for various privacy/control reasons.

Patch-set: 2