Update patch set 2
Patch Set 2:

The mailing list has public archives.  A private G+ group doesn't.  Main issue is getting people there in the first place.  :)

I was too busy with Find7 this week to take care of other stuff, I'm going to talk with the rest of the team about how we want to go forward with this.

Patch-set: 2